Fixed dist_start datastream for groups 4 and 5
This fixed the bug where individuals were failing to settle if their migrations were too short. However, this also introduced a massive assemblage of errors throughout the rest of the model. I could either fuck around with the design matrix for a while to see if I could get the MLE to estimate dist_start distributions for 4 or 5 simultaneously, or I could just go with the 4-state-model, which I know already works. I currently intend to do the latter. This will likely be the last commit for this branch.