Notes on current issues
No changes, but I did a semi-comprehensive sweep through the male results. Extensive issues with stopovers conflating with the pre-mig state, as well as other various issues down the line. At least one post-mig classification that should have been a stopover. This is going to need a dist_start variable to fix, at least, or resume the original starting ruleset.