Breakthrough: limited mig steps to > 500m
Fixed issue with segregation of stopover points with pre/post mig. The resulting model is pretty good, but a little more trigger-happy than I'd prefer. Some birds are being classified as post mig despite transmitter death in early April, for example (and while most are probably correct, some definitely aren't).
Going to run a 4-state with the same fixes to see if it's as trigger-happy. Getting close here...