CHECKPOINT: all models finalized
Added lat back into the fall model, and revised the spring models to use a similar framework to the fall model. Also did lots of cleaning and prepped the accuracy assessments.  At this point we're ready for final proofs of excluded birds/points, accuracy assessments, and integration into a single file. The only change that I think is likely to be made to these models at this point is probabilistic assignments for initial states.