LDM changes to base model
Note- this is not including a full set of integrated designations for the base model, which I'm going to avoid doing for now as the sheer number of birds which don't settle are causing bugs. However, I do have a list of all LDMs which don't occur in the base model and vice versa.